reorganized random conditional files
stoppit kelvin
